+++ src/components/NavBar/buttons/MenuBtn.js
@@ -1,7 +1,16 @@
 import styled from 'styled-components';
-import closeIcon from '../../images/close.svg';
-import userIcon from '../../images/user.svg';
-import { Container } from './Container.styled';
+import closeIcon from '../../../images/close.svg';
+import openIcon from '../../../images/menu.svg';
+
+const Container = styled.div`
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  @media (min-width: 768px) {
+    display: none;
+  }
+`;
 
 const OpenMenu = styled.img`
   margin-right: -24px;
@@ -15,13 +24,13 @@ const CloseMenu = styled.img`
   transform: rotate(${({ opened }) => (opened ? -90 : 0)}deg);
 `;
 
-const SearchBtn = props => {
+const MenuBtn = props => {
   return (
     <Container onClick={props.onClick}>
-      <OpenMenu src={userIcon} alt='' opened={props.opened} />
+      <OpenMenu src={openIcon} alt='' opened={props.opened} />
       <CloseMenu src={closeIcon} alt='' opened={props.opened} />
     </Container>
   );
 };
 
-export default SearchBtn;
+export default MenuBtn;
