+++ src/components/NavBar/buttons/UserBtn.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
-import closeIcon from '../../images/close.svg';
-import openIcon from '../../images/search.svg';
-import { Container } from './Container.styled';
+import closeIcon from '../../../images/close.svg';
+import userIcon from '../../../images/user.svg';
+import { Container } from '../Container.styled';
 
 const OpenMenu = styled.img`
   margin-right: -24px;
@@ -18,7 +18,7 @@ const CloseMenu = styled.img`
 const SearchBtn = props => {
   return (
     <Container onClick={props.onClick}>
-      <OpenMenu src={openIcon} alt='' opened={props.opened} />
+      <OpenMenu src={userIcon} alt='' opened={props.opened} />
       <CloseMenu src={closeIcon} alt='' opened={props.opened} />
     </Container>
   );
