+++ src/components/NavBar/sliders/CartSlider.js
