+++ src/components/NavBar/sliders/MenuSlider.js
